# docs/api/edgechat-v1.openapi.yaml
openapi: 3.1.0
info:
  title: EdgeChat Mobile API
  version: 1.0.0
  description: Stable HTTP contract shared by native clients and EdgeChat servers.
servers:
  - url: https://chat.example.com/api/v1
security:
  - bearerAuth: []
paths:
...
  /channels:
    get:
      operationId: listChannels
      responses:
        "200":
          description: Visible channels
          content:
            application/json:
              schema:
                type: object
                required: [channels, publicChannels, privateChannels]
                properties:
                  channels: { type: array, items: { $ref: "#/components/schemas/Channel" } }
                  publicChannels: { type: array, items: { $ref: "#/components/schemas/Channel" } }
                  privateChannels: { type: array, items: { $ref: "#/components/schemas/Channel" } }
    post:
      operationId: createChannel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, kind]
              properties:
                name: { type: string }
                description: { type: string }
                kind: { type: string, enum: [public, private] }
                memberUserIds: { type: array, items: { type: integer } }
      responses:
        "200":
          description: Created channel
          content:
            application/json:
              schema:
                type: object
                required: [channel]
                properties:
                  channel: { $ref: "#/components/schemas/Channel" }
  /channels/{channelId}/join:
    post:
      operationId: joinPublicChannel
      parameters:
        - $ref: "#/components/parameters/ChannelId"
      responses:
        "200": { $ref: "#/components/responses/Ok" }
  /channels/{channelId}/members:
    get:
      operationId: listChannelMembers
      parameters:
        - $ref: "#/components/parameters/ChannelId"
      responses:
        "200":
          description: Channel members
          content:
            application/json:
              schema:
                type: object
                required: [room, members]
                properties:
                  room: { $ref: "#/components/schemas/Channel" }
                  members: { type: array, items: { $ref: "#/components/schemas/Member" } }
  /channels/{channelId}/invite:
    post:
      operationId: inviteChannelMembers
      parameters:
        - $ref: "#/components/parameters/ChannelId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userIds]
              properties:
                userIds: { type: array, items: { type: integer } }
      responses:
        "200": { $ref: "#/components/responses/MembersMutation" }
  /channels/{channelId}/members/{userId}:
    delete:
      operationId: removeChannelMember
      parameters:
        - $ref: "#/components/parameters/ChannelId"
        - name: userId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200": { $ref: "#/components/responses/MembersMutation" }
  /channels/{channelId}:
    patch:
      operationId: updateChannel
      parameters:
        - $ref: "#/components/parameters/ChannelId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                avatarKey: { type: [string, "null"] }
      responses:
        "200":
          description: Updated channel or no-op result
    delete:
      operationId: deleteChannel
      parameters:
        - $ref: "#/components/parameters/ChannelId"
      responses:
        "200": { $ref: "#/components/responses/Ok" }
  /dm/open:
    post:
      operationId: openDirectMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: integer }
      responses:
        "200":
          description: Direct conversation
...
  /rooms/{kind}/{roomId}/read:
    post:
      operationId: markRoomRead
      parameters:
        - $ref: "#/components/parameters/RoomKind"
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                messageId: { type: [integer, "null"] }
      responses:
        "200":
          description: Read cursor
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ChannelId:
      name: channelId
      in: path
      required: true
      schema: { type: integer }
    RoomKind:
      name: kind
      in: path
      required: true
      schema: { type: string, enum: [public, private, dm] }
    RoomId:
      name: roomId
      in: path
      required: true
      schema: { type: integer }
  responses:
...
    Ok:
      description: Successful command
      content:
        application/json:
          schema:
            type: object
            required: [ok]
            properties:
              ok: { type: boolean, const: true }
...
    MembersMutation:
      description: Updated member list
      content:
        application/json:
          schema:
            type: object
            required: [ok, members]
            properties:
              ok: { type: boolean }
              members: { type: array, items: { $ref: "#/components/schemas/Member" } }
  schemas:
...
    Channel:
      type: object
      required: [id, name, kind]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        avatarKey: { type: string }
        avatarUrl: { type: string }
        kind: { type: string, enum: [public, private, dm] }
        isGeneral: { type: boolean }
        ownerDisplayName: { type: string }
        isMember: { type: boolean }
        myRole: { type: string }
        canManage: { type: boolean }
        memberCount: { type: integer }
        lastMessageAt: { type: [string, "null"] }
        unreadCount: { type: integer }
        mentionUnreadCount:
          type: integer
          description: Legacy field name for Telegram-style direct attention unread messages, including mentions and replies.
    Member:
      type: object
      required: [id, username, displayName, avatarUrl, role]
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string }
        role: { type: string, enum: [owner, member] }
        joinedAt: { type: string }
...